Implemented progressive rate strategy. Changed return type of calculateTime in order to correctly calculate the second case of progressive rate.

+++ src/paystation/domain/AlternatingRateStrategy.java
@@ -3,7 +3,7 @@ package paystation.domain;
 public class AlternatingRateStrategy implements RateStrategy{
 
 	@Override
-	public int calculateTime(int moneyInserted) {
+	public double calculateTime(int moneyInserted) {
 		// TODO Abraham
 		return 0;
 	}
